Change to Source/Program monitor.
Now, program monitor reflects a playlist if only a playlist project or
the timeline if a multitrack project. Double-clicking a playlist item
opens the shot in the source monitor. This makes it more convenient to
use as a bin for a timeline project. Shift-double-click can seek to a
shot in the playlist.