Make it easier to adjust image duration on timeline.
Previously, the duration of an image could not be stretched longer than
the image duration in Properties, which defaults to 4 seconds. It is not
always obvious or convenient to change image duration in Properties.
Instead, let image duration in Properties adjust the out point, and let
image duration be stretched up to 10 minutes in Source player or on
timeline.