Improve switching between filters.
Before this change, when selecting a text filter while a different text filter was already selected would result in the textbox not being refreshed with the new value. This change ensures that the filter pane is cleared before a new filter pane is loaded - even if the new filter is the same type as the previous filter.