Make playlist update button use index set on Source.
Now, the currently selected playlist item does not need to match what is
open in the Source player. Rather, the update button will use the index
that was set on the producer in the Source player and use that. This
makes it more convenient and natural. Also, the Update playlist menu
item was changed to Replace and can be used to replace any playlist item
with any clip opened in the Source viewer.