Modify volume behavior.
Producers can producer audio even when paused or seeking. This way, the affect
of audio filters can be visualized in the audio waveform scope when the player
is paused.

Also, the volume control not only applies to the consumer, and not the rest of
the application. This way, muting or changing the volume in the player does not
affect the waveform and VU meters.

Consumer audio is muted when the player is paused to avoid annoying blips and
pops when the consumer is refreshed due to filter parameter changes.