Dependency module version tracking
The exact revisions of the dependency modules are now recorded in
build-shotcut.sh, so that we can see exactly what commit went into what
release, and what changed between them.

See the comment in build-shotcut.sh for the syntax. To update these listings,
just run update-revs.sh and it will fetch, compare, update the file and commit
(but not push, of course)