Fix timeline thumnail deadlock on Win64.
The deadlock occurs when creating a Mlt::Producer. It did not matter
what service was requested. Even a color producer fails. Returning an
image created and filled by Qt also works fine. The problem is some
strict combination of ForceAsynchronousImageLoading and Mlt::Producer.
After removing the flag, doing the work of requestImage in a separate
thread and then waiting for the result succeeds proving that just
working with MLT in a separate thread is not a problem (as demonstrated
elsewhere - such as getting audio levels).