Webvfx filter ui: custom ui
Custom UI for your html filter can now be included in the filter UI by
including it next to the html file named "[htmlfilename]_ui.qml"

For convenience, a SimplePropertyUI item is also included for when all you
want to is to set a list of properties for your item. Example:

myfilter.html:

<html><body>blabla</body></html>

myfilter_ui.qml:

import QtQuick 2.1
import Shotcut.Controls 1.0
SimplePropertyUI {
    properties: ["propertyA", "propertyB"]
}

This will have the filter UI present two text fields that set propertyA and
propertyB on the clip.