Add dynamic OpenGL for Windows.
Dynamic OpenGL is the ability to load an OpenGL implementation at run-
time. Qt 5.5 includes support for desktop OpenGL 2+, ANGLE for Direct3D,
and Mesa+llvmpipe for software rendering. This is user-selectable in the
Settings menu, which defaults to Automatic. GPU processing only works
with OpenGL, whether explicit or selected automatically. ANGLE and Mesa
do not support multi-threading.