Remove SharedFrame::get_image(args).
The new get_image() method with args is dangerous because it can be used
to change the image format, which violates the const nature of
SharedFrame. Instead, when needed, the FrameRender in GLWidget converts
the image to the format it needs on its thread before making and sharing
the SharedFrame.