Check XML for problems and offer/try to repair.
In this first pass, it is only looking for mixed usage of comma and
period as the decimal mark. Upon repair, it changes values to be the
same as the system numeric locale and removes LC_NUMERIC attribute from
the root mlt element. It does not yet handle all numeric values because
there are many to identify, tries to be conservative, and does not
attempt to determine if every string is numeric. It only checks in, out,
length, and geometry.