Refactor Database to keep sql operations in single thread
All database operations will now post a job request for the database thread,
then block until it is completed. Testing shows that in practice there is
rarely waiting going on due to multiple jobs in the queue.
The most expensive operation of all: writing to disk, has been delayed until 5
secs after the last database request, and all operations are grouped in an
implicit transactions. This improves thumbnail loading time greatly as the disk
is not written to before thumbnails are ready and on screen, and more
importantly, only written to once (not one time for every thumbnail in the
timeline).