Move canvas painting to QQuickPaintedItem
This commit takes a different approach on the qt >=5.5.0 canvas item bug
that has proven very hard to workaround reliably. Instead of doing the drawing
in canvas with javascript, the drawing code is moved to a small set of
QQuickPaintedItem classes that do the same drawing with QPainter (which
provides a very similar api).

In order to speed up and avoid detaching, this commit also includes a change to
how audio levels are passed along - now the pointer to the levels data is made
available through a QVariant, and the waveform painter will access it using
in/out points and read directly from that data without deep copying.