Use shotcut:hash to index thumbnails and waveforms.
Sometimes, the resource (file path) is a relative path - when it is in
the same folder as or a subfolder of the project. In these cases, it is
much more likely that a file is named the same (e.g., screencast.mkv or
audio.wav) resulting in the same db key. This approach uses an existing
hash of the actual file contents to give unique results and correct
thumbnails and waveforms.