Add noupgrade config and command line option.
Either add "noupgrade" = "true" in the config/setting file, or add "--
noupgrade" to the command line to suppress the display of the startup
message to "Click here to check for a new version of Shotcut." This also
removes the Help > Upgrade menu item. Some school lab administrators
have requested this option.