Fix #361 JACK client connects itself.
Now, it automatically connects the output ports to system playback, but
it does not connect the system capture ports as this client is for the
Shotcut player and not recording (input ports still there for corner use
cases).

This also fixes #463 by using a new MLT "jack" filter to control the name
of the JACK client. It uses "Shotcut player" because "Shotcut" is used
for capturing from JACK using File > Open Other > JACK Audio.