Fix track name not editable regression in 78d3724.
This is an unfortunate workaround that prevents being able to edit the
track name when the Timeline dock is floating. I could not come up with
anything better using sendEvent in the KeyPress handler or an event
filter on the QQuickWidget. Things might be different with a newer
version of Qt.