Fix AvformatProducerWidget inaccurate with timeline clips.
The MLT avformat producer needs to decode a frame in order to get some
info like scan mode and field order. However, this widget was simply
trying to read the producer after the next frame displayed. But on
Timeline, the play head might not be in the selected clip and thus not
triggered this selected clip's avformat producer to decode a frame. The
fix is to really decode a frame from this producer (asynchronously).