Only adjust filter when trimming clip if filter is not trimmed.
The filter is only adjusted if it's in point is the same as the clip's
in point. Likewise for the out point. Thus, the filter is automatically
extended along with the clip as long as it is not trimmed. If the filter
was trimmed, treat it like the filter's region of time is absolutely
positioned (think: correction filter use cases).