Fix many frames dropped on start of playback from pause.
This happens when Realtime is enabled and GPU effects disabled. This is
related to #571 and something bothering me for a while. This reverts
many of the changes in ab3ce28d. I never saw the problem it describes in
today's testing, and the part about "consumer would reload a bunch of
the same frame" is minor and not as important as the user experience.