Add HTML template for WebVfx JS with Web Animations polyfill.
If you choose Filter > Overlay HTML > Use WebVfx JavaScript extension >
New..., we would intialize a broken empty doc with no WebVfx
integration. With this change not only is basic WebVfx JS code included
to unbreak it, but it also contains the Web Animations (https://
www.w3.org/TR/web-animations-1/) polyfill with support for properly
seeking on CSS and Web animations. Furthermore, it contains the basic
example code to animate the body of the document in the style of
scrolling credits.
P.S. For some reason, the web-animations.js is not loading properly as a
Qt resource and needed to be included directly in the document.