Fix #615 by removing desktop-helper from snap.
desktop-helper is affecting environment variables and just getting in
the way since we are not trying to use the snap-installed Qt libs
anyways. This basically just makes the snap a variant of the portable
tar. I hope we can soon make a proper custom snap build ala the flatpak
to replace this whole approach and also run with confinement.