Remove composite toggle from track head.
This is moved into the track Propertie's Blending mode drop-down, which
can now be set to None to disable compositing (blending). It is not
available on the bottom video track.
Also added Ctrl+Alt+B keyboard shortcut to toggle blending. Alt was
added as a modifier to make it more intentional. I do not want people
using the B overwrite shortcut to accidentally use Ctrl+B and toggle the
blending.