Fix audio levels rebuilding at shutdown.
This was causing the process to linger for potentially long periods of
time and loading the computer. This change also disables generating
levels if the previous database write fails (often it becomes locked by
another shotcut process) because that is a potentially heavy task whose
result is unable to be cached. Also, prevent generating waveforms for
known silent producers.