Make GPU Effects hidden and discouraged.
Hidden means it is no longer in the Settings menu, and you must edit the
configuration file/registry to turn it on (except when switching it on
for a project using GPU effects). Discouraged means you must acknowledge
a warning dialog every time the app starts with GPU effects enabled.