Fix MLT locale of qmelt jobs.
Fixes #745 and #713

MLT applies the system locale by default, which we can control in the
current process mainly using setlocale(). However, we need to set an
environment variable to affect the child process (variable already needed
in Windows by the xml consumer's call to getlocale).

Next, we need to inform users of Export > Other how to enter numeric
values depending on the MLT locale state of the current project. And we
need to conform encode presets' floating point numeric values.