Fix save XML backup fails if removing temp file fails.
The temp file can fail if something like Dropbox puts a lock on it:
https://forum.shotcut.org/t/error-saving-project-mlt-with-0-bytes/11656/

The solution here does not try to remove the temp file but instead copy
data from the existing file into it and then delete the existing file.
Likewise, something can put a lock on the XML consumer temp file
preventing to rename it to the destination. A simple retry mechanism
resolves it for me on Dropbox, but it depends on the speed of whatever
has the lock to release it.