Fix filters added after transitions.
Previously, adding a filter to a clip involved in a transition did not
extend that filter into the time range of the transition, and the
Keyframes filter trimming did not make it possible to extend it either.
Now, it works as expected. Also, there were some general fixes for
filter in/out with undo.

https://forum.shotcut.org/t/error-in-transition-between-pictures-when-
working-with-keyframes/11540