remove the YADIF options from Settings > Deinterlacer
This is a hopefully temporary workaround, but I have not been able to
find the root cause in MLT. YADIF is crashing on multiple threads
whenever a mlt_transition is involved. This includes things such as
dynamictext and affine filters. In Shotcut, when frame-stepping using
keyboard or scrubbing with the mouse, multiple threads are also in use.
This does not address export with parallel processing on, but if that
fails it automatically restarts without parallel processing and almost
always succeeds thereafter.