Make the build reproducible
Whilst working on the Reproducible Builds effort [0] we noticed that
shotcut could not be built reproducibly.

This is because it uses the current build date for the SHOTCUT_VERSION
variable. A patch is attached that bases this value on
SOURCE_DATE_EPOCH [1] instead.

This was originally filed in Debian as #949817 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/949817