fix shift+mousewheel on macOS, workaround mousewheel to scroll
Qt on macOS is making the mouse wheel look very similar to track pad,
which makes the logic go into the track pad block and then interpreted
in a way that is incompatible with the track pad behavior. The
workaround is to make the modifiers strict, allow no modifiers for the
track pad, and let any undefined modifier or combinations fall through
to horizontal scroll.