fix OpenGL on Linux build and sync lib exclusions with AppImage
I noticed that GPU effects stopped working, and on on my system it was
allocating a GL 2.0 context instead of the 4.6 I usually see in my
development workspace. This could potentially be a problem with Quick
rendering and video preview as well. I was suspicious of libGLdispatch
in the bundled libraries and saw AppImage had excluded it. Removing it
solved the problem.