fix deadlock dragging multiple files to timeline
A mutex is not needed because this is UI code that must be called on
the same thread. Secondly, the LongUiTask cancel invoked itself instead
of QProgressDialog::cancel(). And third, the long task progress dialog
was always canceled by the AvFormatProducerWidget and not only when a
dialog would be displayed.