fix qml menus not showing
So far this is only known to affect Windows when using multiple screens
with different display scales in Windows Settings. The workaround is to
change the display scale rounding policy to RoundPreferFloor, i.e. half
round down such that 1.5 is treated at 1, but 1.75 as 2. This still does
not cover all cases in my testing. 200% on the primary screen and 100%
on the secondary, for example, is not working on my Windows workstation.

This changes adds a new command line option `--
QT_SCALE_FACTOR_ROUNDING_POLICY` to set the rounding policy (easier for
windows easier than environment variable). The default is
RoundPreferFloor. This choice should also help users complaining that it
is too big now on a 150% screen

https://forum.shotcut.org/t/display-a-menu-of-additional-actions-doesnt-
open/22314
https://forum.shotcut.org/t/ver-20-10-31-right-click-mouse-menu-not-
working/22275