fix quick controls 2 toolbars accepting focus
This is confusing because tab to change focus is not necessarily
expected on a toolbar, and this leads to accidental activation of
toolbar items when using space key to toggle play/pause. (The same is
not true for forms like filter parameters, properties, export, etc. where
tabbing is expected and supported.)