disable focus on combobox because due to bad keyboard control
https://forum.shotcut.org/t/filter-preset-dropdown-changes-erroneously/
24296

Even without Shotcut toggling playback on the space bar, the fact that
it automatically selects whatever happens to be under the mouse when you
open the combo using the keyboard is bad behavior. Changing
`hoverEnabled` based on `focusReason` seemed promising, but the Preset
combo reapplies the preset when you reselect it, which is a feature, and
it is too easy to close the combo with space - perhaps in quick response
to playback changing. I was unable to find any other way to prevent the
space bar from also toggling playback while operating the combo
including things like `FiltersDock::event()` and `Keys.onSpacePressed`.