fastForward_Speed_decrease
In addition to your change "Improved the playback speeds of fast forward and rewind to not be so fast and more usable."
I think you should reduce the speed increase of fastforward button since it can be accelerated by pressing on the button again.
This will permit for instance to still understand people talking in the video for instance. Since it is quite impsoobile at the double speed.