add repair for near future version of project
Soon there will be a version based on MLT v7 with chain and link
elements for time remapping. This adds the ability to load these future
projects by stripping the time remap effect and converting a chain to a
regular producer. It also detects if a project was made with a MLT
version greater than the next major version and show a meaningful dialog
about incompatibility due to versions.