Shift keyframes for all filters upon split or trim
When the in point of a producer is changed due to split or trim,
adjust the keyframes for all filters attached to that producer.

For simple keyframes, keyframes are always relative to the start and
end of the filter. So the 3rd and 4th keyframe need to be moved if
the out point changes.

For advanced keyframes, all keyframes are relative to the start of
the producer. So all keyframes need to be moved if the in point
changes.

Issue reported here:
https://forum.shotcut.org/t/inconsistent-handling-of-keyframes-when-splitting-a-clip/27404