Make the Deband filter Undo buttons match the default preset (#1133)
* Make the Undo values match the default preset

"FFmpeg default values" used to be the default preset before being renamed to "Average strength".  Later, changed the default preset to be "Full to limited range".  This commit updates the default values to match "Full to limited range" so that the Undo buttons will appear to match the default preset.

* Clearly define values for each preset

Make it easy for future edits to change values of presets by defining values now for each preset rather than daisy chaining.