use filesystem instead of sqlite for thumbnails (#1141)
* use filesystem instead of sqlite for thumbnails

* do not remove the old db

Removing the db makes it such that running an older version does not
utilize the old cache. Sure, the older version will not be able to
utilize new cache data, but it is a bummer to wait for waveforms when
running an older version against existing media. Also, if a person
downgrades.

* do not generate audio levels if no audio track

* run thumbnail conversion at startup