Fix crash when adding 2 variable frame rate clips
As reported here:
https://forum.shotcut.org/t/adding-at-least-2-variable-framerate-clips-from-playlist-to-timeline-crashes-shotcut/29411

The "convert" dialog is created on the stack in the avformat producer
panel. When adding items from the playlist to the timeline, each item
triggers the panel to load. When a new clip loads in the panel, the
previous panel is deleted and tries to delete the dialog if it was
created. Attempting to delete a dialog that was created on the stack
causes a crash.

This change only triggers events on the last item so that the panel
will not open and delete for each item being added.