Improve audio distortion when reversing a clip.
Use alac and mov for all intermediate files for reverse. This
combination results in the best audio syncing and the least
artifacts when reversing a clip.

This is not a definitive fix. Some further improvements will
need to be made in avformat producer to improve audio seeking.

As reported here:
https://forum.shotcut.org/t/reversing-a-clip-on-linux-causes-the-audio-to-crackle/31713