Workaround to fix selection after moving a clip
As reported here:
https://forum.shotcut.org/t/track-rearrangement/27259/5

It seems there is a race condition between the multitrack model
updating and the selection model updating in the QML. The result
is that when moving tracks with selected clips, sometimes clips
show as selected but they are not selected.

This workaround reloads the entire timeline QML and then resets
the selection.

This commit may be reverted if this workaround is too slow with
large projects.