improve text size for Notes
On Windows and macOS Qt applications use different font sizes for
different classes of widgets. This is even more noticeable on Windows
with display scaling over 100%, which is common. As a result, the Notes
text size is rather small, which can be bothersome with any non-short
text length. This change makes the text size the same as what appears in
the menus, which looks better.

Also, changed the tab size to use a non-deprecated method and use a
character that does not contract with kerning.