fix video track blend broken after moving a track
https://forum.shotcut.org/t/image-losing-transparency/34056

This happens because the track types become interleaved at the MLT level
as they are added out-of-order. What is interesting is that the blend
mode property "1" is being tracked for b_track changes, but the b_track
property itself was not. I believe that is because it believed the track
types will not change and the MLT video transitions can stay in place
and simply alter the blend mode as needed. However, the track types can
change as they are moved.