Fix The slider in the UI is not updated when a keyframe is deleted (#1361)
As reported here:
https://forum.shotcut.org/t/the-slider-in-the-ui-is-not-updated-when-a-keyframe-is-deleted/36150

QmlFilter has two signals with the same name:
 * changed()
 * changed(QString)

The QML Connections item can not bind to a signal when it is overloaded
by another signal with the same name but different parameters.
As a result, the onChanged() signal never fires in ui.qml.

This change merges the two similar signals into one. If the parameter
name is not known, or does not apply, pass an empty string.