Fix file save dialog does not open in Ubuntu
Using Ubuntu 18.04 and using the transcode dialog. If the dialog
is opened using the Convert button, the File Save As dialog works
properly. However, if the transcode dialog is triggered by the
Time Remap filter (due to using a source with b-frames), the
File Save As dialog does not pop up when the user clicks "OK" and
Shotcut hangs waiting for the dialog to appear. The dialog is not
"under" the application. It does not appear at all.

This behavior exists for both Qt5.15 and Qt6.4.1.

This problem does not occur on Windows.