use system palette
Since we say we are limited to using the system color theme, then we
should actually use it. This might require a little tweaking as I have
done here on macOS to avoid some ugliness since we are not using native
control _styles_ - because those are not actually aligned very well with
the modern OS control styles and look very ugly!