switch to yuv420p as the non-GPU-mode MLT format
This generally improves preview performance by less conversions as most
sources are yuv420p, and the video widgets use yuv420p. And most CPU
effects use rgb(a); so, before it was converting to yuv422 and then to
yuv420p. GPU mode uses rgba because that is the movit output, and when
you add CPU filters, again most are rgb(a).