fix playback sound on Windows after SDL upgrade
SDL in the Windows build was upgraded from version 2.0.5 to version
2.28.1 to fix a bug where sound is not follow changes to the default
output device in Windows

https://forum.shotcut.org/t/sound-into-headphones/33298
https://forum.shotcut.org/t/help-shotcut-wont-switch-audio-playback-
devices/22038
https://forum.shotcut.org/t/audio-output-does-not-change-with-computer-
on-windows/14549
https://forum.shotcut.org/t/shotcut-will-not-play-audio-through-
bluetooth-headphones/11112

SDL version 2.0.6 fixes this but also introduced the default driver
"wasapi" that creates a buzzing or crackling sound when the system is
configured for 5.1. It also plays stereo as multi-channel stereo like
many surround receivers offer but even when the receiver is not
configured as such! "directsound" fixes those issues, but a different
kind of distortion appears after changing output devices. "winmm" fixes
that, but it does not work when opened with more than 2 channels. What a
mess.