fix key behavior of Actions and Shortcut on macOS
The macOS item activation shortcut is command-O, which is not only
completely not-obvious to a veteran mac user such as myself (I only
expect that in Finder) but instead triggers File > Open! Meanwhile,
Enter, which is supposed to trigger the item, instead closes the dialog
without triggering. Also, it would not consistently give the search field
focus on showing the dialog.