Notify the service when the keyframe time changes
When adding or removing a keyframe, the API for the service is used
and the service will be notified if it is listening to property changes.

However if the keyframe is changed (type or frame), the Animation API
is used and the service is never notified that the property is changed.
Notify the "property-changed" event from the application so that the
service knows about it if it is listening.