Fix crash when undo triming transition to a new transition. (#1484)
As reported here:
https://forum.shotcut.org/t/using-undo-after-dragging-transitions-sometimes-cause-crashes/41096

Do not allow the user to trim the transtion past "remove". When they hit the
end of the transition, they can either commit, or untrim. Before this change,
they could continue to trim and add a new transition. But the new transition
command does not know about the removed transition change and that can lead
to a crash when undoing.

Also, fix transition diasppearing when it is trimmed out and then back in
without releasing the mouse.