Fix filter undo breaks after a timeline move
Make all timeline clips store and search for UUID on the clip parent.
The filter commands use the clip parent becuase they do not have
access to the cut. All other operations need to also keep and restore
the UUID on the parent so that filter commands will continue to work.