Fix selection becomes inconsistent upon undo/redo
Sometimes the UI state would be incorrect. For example, a menu entry
might be disabled when it should be enabled. This is becasue the
selection change signal was not sent.

Replace the selection blocker with a selection silencer. Previously,
the selection would change, but if the blocker was in place, the
change would not be reported. The silencer will ensure that the change
is eventually reported.