Move isMemoryLow to mainwindow.cpp
MainWindow is the only consumer of the isMemoryLow function, so the function does not need to be part of the Util class's interface. The Util class is currently kind of a junk drawer with low cohesion among the class functions, so it may be helpful to reduce the surface area of Util.