separate Projects from Recent
The goal is to prevent projects unexpectedly disappearing in the Shotcut
UI because a lot of files have been opened. A number of users have
expressed alarm or concern for missing projects. Thus, projects are
saved indefinitely and with no limit. But also still listed in Recent
because Recent represents user file activity regardless of type.

Projects are saved under config key `projects` but still in the
`recent.ini` file in the app data dir.