add status/feedback to Files
The toolbar now shows an item count updates periodically on very full
directories to show some progress and activity.

The Locations combo now shows the current path, which is nice to have
when the folders tree is closed. Plus, it can be copied, and pasted into
to change to a directory.