remove JACK audio capture and deprecated screen capture
x11grab and gdigrab were hidden a long time ago as inefficient and
troublesome. Moreover, X11 is on its way out with Wayland. Time to
remove these as well as the hard dependency of linking FFmpeg with
libjack to enable a rarely used JACK audio capture for which there are
better tools. Besides, JACK has been integrated with PipeWire now, which
is available through our existing PulseAudio API integration.