don't auto-laumch Glaxnimate on Add to Timeline
When doing Open Other > Animation > Add to Timeline, Glaxnimate launches
without Shotcut video as the canvas background. This is because the
producer loaded in the new timeline clip is different than the one
constructed by the dialog. Instead of adding much complexity or ugliness
to the code, simply do not launch in this case. The UI still switches to
the Properties dialog. Then, it is fairly obvious to click Edit to
launch from there. At least, one should know how to get back to Edit or
be forced to learn it.