Add File > New submenu and Timeline > New Generator
It has been weird to use **Open Other** to access generators. This
change separates generators from **Open Other**, which is really only
appropriate for network and devices. The generators are put into a new
**File > New** sub-menu. The Open Other main toolbar button changed its
name to **New Generator** and only lists generators. This button is
honestly no longer needed on the main toolbar but kept to prevent change
from confusuing some. Furthermore, it is also strange to remember to
click **Add to Timeline** when that is the primary use case. Therefore,
added a **New Generator** icon to the **Timeline** toolbar to streamline
that (**OK** instead of **Add to Timeline** and **Open** buttons).