Fix #1690 1 frame fade out is mute/black
There was a miniumum duration of 2 frames in the filter's Duration
field, but it could be defeated by not focusing out within the UI (click
somewhere outside the Filters panel). And that was also confusing to
some. So, now that is 1 in the UI and minimum 2 (1=2) on the engine's
keyframe specification.