Fix download of whispermodel.qml
Qt may log "Incompatible version of OpenSSL (built with OpenSSL 1.x,
runtime version is >= 3.x)" on Qt 6.4 that is in Linux AppImage,
portable, and Snap. This unfortunately does not fix the actual model
downloading, which is only over HTTPS. But at least now it does not
crash. Users affected by this will need to manually download a model
still and configure it manually in the dialog.