Revert custom OpenCV build command (#1700)
An older OpenCV branch lacked the definition of M_PI in
carotene's definitions.hpp. To work around this, a custom
build_opencv function was introduced to patch the file
and handle the build manually.

With OpenCV 4.x, this definition is now included by default
as of commit ce1d840adfe68504fd9b35f1568286e1911599c9. The
workaround is no longer necessary.

This patch removes the build_opencv function and restores
the original OpenCV build command.

Signed-off-by: Thirumalai Nagalingam <thirumalaicse2020@gmail.com>