Use ScreenCapture for Screen Snapshot on Windows
This is a better integration since we get back a file name and can open
it. It still supports Rectangle and Fullscreen options. Ported
ScreenCaptureJob to Windows using ffmpeg, but this is still not used for
now due to poor performance of gdigrab even with a hardware encoder
(tested 4K60). Actually, x264 drops much less frames such that 1080p30
is tolerable with less than half total CPU usage on my system. However,
it does not scale up well to higher resolutions and frame rates, nearly
using all CPU, and that will conflict with being able to record some
activities that also need to make good use of CPU or not have terrible
latency.